Separate out system.py
